# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: upcontrol API
  version: 1.0.0
  description: |
    The front-facing contract for upcontrol. Generated into a strict Go server
    (oapi-codegen) and into TS types for `front/src/lib/api.d.ts`. Response
    shapes are pinned to `front/src/lib/mockData.ts` — a contract test
    (build-plan §2.7) checks each against golden JSON taken from the mocks.

    Rules from the plan (§4.4):
      - Version in the path (/v1), not a header.
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
        text reason. No `disabled`, no `navigate('/plan')`.
      - Pagination only on unbounded lists: incidents, logs. Monitors, channels
        and recipients are plan-bounded and returned whole.
      - `null` is never used for "no value": the field is omitted. "Zero is
        silence" — an absent field does not render.
      - All timestamps are RFC 3339 UTC.
      - `POST /i` and `POST /hooks/{provider}` are `x-hand-written: true`: they
        exist here for the /docs/api page but are implemented by hand (the
        generator's validation is forbidden on /i by §4.5).
servers:
  - url: http://localhost:5173
    description: local dev (front dev server proxies to ucapi)

# ---------------------------------------------------------------------------
# Common building blocks
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: uc_session
    apiKey:
      type: apiKey
      in: header
      name: X-Upcontrol-Key
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No session, or the key is not ours.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Resource does not exist or belongs to another tenant.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: Rate limited (always with Retry-After).
      headers:
        Retry-After:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "plan_limit_exceeded" }
            message: { type: string }
            upgrade:
              type: object
              description: Present only on 402. Drives the client's upgrade prompt.
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "Free allows 3 HTTP checks."
                highlight:
                  type: string
                  description: Optional row to highlight on the pricing table.

    # --- account ------------------------------------------------------------
...
    ProjectMeta:
      # The installer-collected project spec (ai-provider plan decision 15b):
      # ONLY these five fields. Never dependency lists, versions, file paths,
      # git remotes, env values or code — `npx upcontrol init` prints the exact
      # spec before sending and `--no-meta` skips the upload entirely.
      type: object
      properties:
        name: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 200 }
        framework: { type: string, maxLength: 200 }
        runtime: { type: string, maxLength: 200 }
        language: { type: string, maxLength: 200 }

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /v1/projects/anonymous:
    post:
      tags: [install]
      security: []
      x-hand-written: true
      summary: |
        Mint an anonymous project (cli/SPEC.md §7.1): tenant + project + API key
        with no person attached, so `npx upcontrol init` works before signup.
        The body carries CLI/agent versions only — no email, no hostname, no
        paths. The key rides the response exactly once; the claim token is the
        one-time proof that later attaches a signed-in person to the tenant.
        Claiming never changes the key. Per-IP throttled.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                cli_version: { type: string }
                agent_version: { type: string }
                platform: { type: string }
                arch: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [projectId, key, claimToken, claimUrl]
                properties:
                  projectId: { type: string }
                  key: { type: string, description: Shown exactly once. }
                  claimToken: { type: string }
                  claimUrl: { type: string }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/claim:
    post:
      tags: [install]
      x-hand-written: true
      summary: |
        Attach the signed-in person to the anonymous tenant the claim token
        names (cli/SPEC.md §7.3). One-time: the token hash is cleared on
        success. Idempotent for an already-claimed member only via membership's
        ON CONFLICT; an unknown or spent token is 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [claimToken]
              properties:
                claimToken: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [claimed]
                properties:
                  claimed: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/install/status:
    get:
      tags: [install]
      security:
        - apiKey: []
      x-hand-written: true
      summary: |
        What `npx upcontrol verify` polls (cli/SPEC.md §8): key-authenticated
        (X-Upcontrol-Key or Bearer), answers whether install_verified has
        arrived inside the visible window, how many lines the window holds, and
        which message names arrived in the last 15 minutes — enough to tell
        "key never worked" from "points placed in code that has not run" from
        "names drifted off the dictionary".
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [verified, lines, recent]
                properties:
                  verified: { type: boolean }
                  verifiedAt: { type: string, format: date-time }
                  lines: { type: integer }
                  recent:
                    type: array
                    items:
                      type: object
                      required: [name, count, lastAt]
                      properties:
                        name: { type: string }
                        count: { type: integer }
                        lastAt: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/install/token:
    post:
      tags: [install]
      x-hand-written: true
      summary: |
        Mint the one-time install token the dashboard's install card embeds in
        `npx upcontrol init --token ...` (front-distribution-alignment.md §1).
        Session-authed; bound to the session's project; TTL 10 minutes;
        single-use. The dashboard never shows a bare `npx upcontrol` — signed
        out of the terminal it would mint an anonymous project and route this
        account's logs past it.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [token, command, expiresAt]
                properties:
                  token: { type: string }
                  command: { type: string }
                  expiresAt: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/install/redeem:
    post:
      tags: [install]
      security: []
      x-hand-written: true
      summary: |
        Burn an install token and issue an ADDITIONAL api_key for its project,
        returning the secret exactly once (the CLI writes it into .env and
        never prints it). Not a rotation - nothing deployed breaks. A spent,
        expired or unknown token all answer the same 404. Per-IP throttled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [key]
                properties:
                  key: { type: string, description: Shown exactly once. }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/project/meta:
    put:
      tags: [install]
      security:
        - apiKey: []
      x-hand-written: true
      summary: |
        The installer's project-spec upload (ai-provider plan decision 15b):
        `npx upcontrol init` PUTs the five optionally-filled fields it printed
        to stdout, key-authenticated exactly like POST /i. Stored after the
        ingest scrubber runs over every value and every newline is stripped
        (the spec renders inside prompt fences); read only as Explain context.
        Unknown keys are dropped, null fields are rejected as the type error
        they are, and a value over 200 characters (counted as sent, before
        redaction expands it) is rejected — never silently truncated. PUT
        REPLACES the whole spec: a field left out of the body is dropped from
        storage, so send every field the spec still carries, which is what
        `init` does on each run.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectMeta" }
      responses:
        "204": { description: Stored }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # --- hand-written (exist here for /docs/api; implemented by hand) --------
...